requirements-win7.txt
opencv-python
pillow
pyinstaller
twain
pywin32
wia_scan
numpy

